requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
sentence-transformers>=2.2.0
scikit-learn>=0.24.2
pinecone-client>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0
langchain>=0.1.0
numpy>=1.19.5
torch>=1.9.0
transformers>=4.11.0